requirements.txt
black
pyglet
PyOpenGl
marshmallow
marshmallow-enum
marshmallow-oneofschema
mypy
numpy
tinyarray
